+++ h3ron-graph/src/io/mod.rs
@@ -1,5 +1,3 @@
-#[cfg(feature = "io_gdal")]
-pub mod gdal;
 #[cfg(feature = "io_osm")]
 pub mod osm;
 #[cfg(feature = "io_serde_util")]